Pull #8797: bump Saxon-HE from 9.9.1-7 to 10.0

+++ src/main/java/com/puppycrawl/tools/checkstyle/xpath/AttributeNode.java
@@ -134,7 +134,7 @@ public class AttributeNode extends AbstractNode {
      * @return {@code AxisIterator} object
      */
     @Override
-    public AxisIterator iterateAxis(byte axisNumber) {
+    public AxisIterator iterateAxis(int axisNumber) {
         throw throwUnsupportedOperationException();
     }
